better test case support on Windows
