+++ tests/test_indexpipelines.py
@@ -1,25 +1,12 @@
 import pyterrier as pt
 import tempfile
 import unittest
-from .base import BaseTestCase
+from .base import TempDirTestCase
 import os
 import pandas as pd
 import shutil
 
-class TestIndexPipelines(BaseTestCase):
-
-    def setUp(self):
-        # Create a temporary directory
-        self.test_dir = tempfile.mkdtemp()
-        print("Created " + self.test_dir)
-
-    def tearDown(self):
-        # Remove the directory after the test
-        print("Deleting " + self.test_dir)
-        try:
-            shutil.rmtree(self.test_dir)
-        except:
-            pass
+class TestIndexPipelines(TempDirTestCase):
 
     def test_add_dup(self):
         def _first(df):
